requirements.txt
flask
flask-cors
flask-smorest
flask-jwt-extended
marshmallow
python-dotenv
flask-socketio
redis
eventlet
flask_sqlalchemy
sqlalchemy
flask-migrate
pymysql